mercurial/loader: Remove the redundant try except clause
It's now properly dealt with in loader-core layer.

+++ swh/loader/mercurial/bundle20_loader.py
@@ -64,6 +64,8 @@ class HgBundle20Loader(SWHStatelessLoader):
         self.temp_directory = self.config['temp_directory']
         self.cache1_size = self.config['cache1_size']
         self.cache2_size = self.config['cache2_size']
+        self.working_directory = None
+        self.bundle_path = None
 
     def cleanup(self):
         """Clean temporary working directory
@@ -123,46 +125,39 @@ class HgBundle20Loader(SWHStatelessLoader):
             directory (str/None): The local directory to load
 
         """
-        self.working_directory = None
-        self.bundle_path = None
         self.branches = {}
         self.tags = []
         self.releases = {}
 
-        try:
-            if not directory:  # remote repository
-                self.working_directory = mkdtemp(
-                    suffix='.tmp',
-                    prefix='swh.loader.mercurial.',
-                    dir=self.temp_directory)
-                os.makedirs(self.working_directory, exist_ok=True)
-                self.hgdir = self.working_directory
+        if not directory:  # remote repository
+            self.working_directory = mkdtemp(
+                suffix='.tmp',
+                prefix='swh.loader.mercurial.',
+                dir=self.temp_directory)
+            os.makedirs(self.working_directory, exist_ok=True)
+            self.hgdir = self.working_directory
 
-                self.log.debug('Cloning %s to %s' % (
-                    self.origin['url'], self.hgdir))
-                hglib.clone(source=self.origin['url'], dest=self.hgdir)
-            else:  # local repository
-                self.working_directory = None
-                self.hgdir = directory
+            self.log.debug('Cloning %s to %s' % (
+                self.origin['url'], self.hgdir))
+            hglib.clone(source=self.origin['url'], dest=self.hgdir)
+        else:  # local repository
+            self.working_directory = None
+            self.hgdir = directory
 
-            self.bundle_path = os.path.join(self.hgdir, self.bundle_filename)
-            self.log.debug('Bundling at %s' % self.bundle_path)
-            with hglib.open(self.hgdir) as repo:
-                self.heads = self.get_heads(repo)
-                repo.bundle(bytes(self.bundle_path, 'utf-8'),
-                            all=True,
-                            type=b'none-v2')
+        self.bundle_path = os.path.join(self.hgdir, self.bundle_filename)
+        self.log.debug('Bundling at %s' % self.bundle_path)
+        with hglib.open(self.hgdir) as repo:
+            self.heads = self.get_heads(repo)
+            repo.bundle(bytes(self.bundle_path, 'utf-8'),
+                        all=True,
+                        type=b'none-v2')
 
-            self.cache_filename1 = os.path.join(
-                self.hgdir, 'swh-cache-1-%s' % (
-                    hex(random.randint(0, 0xffffff))[2:], ))
-            self.cache_filename2 = os.path.join(
-                self.hgdir, 'swh-cache-2-%s' % (
-                    hex(random.randint(0, 0xffffff))[2:], ))
-
-        except Exception:
-            self.cleanup()
-            raise
+        self.cache_filename1 = os.path.join(
+            self.hgdir, 'swh-cache-1-%s' % (
+                hex(random.randint(0, 0xffffff))[2:], ))
+        self.cache_filename2 = os.path.join(
+            self.hgdir, 'swh-cache-2-%s' % (
+                hex(random.randint(0, 0xffffff))[2:], ))
 
         try:
             self.br = Bundle20Reader(bundlefile=self.bundle_path,
@@ -486,6 +481,7 @@ class HgArchiveBundle20Loader(HgBundle20Loader):
     def __init__(self):
         super().__init__(
             logging_class='swh.loader.mercurial.HgArchiveBundle20Loader')
+        self.temp_dir = None
 
     def prepare(self, *, origin_url, archive_path, visit_date):
         self.temp_dir = tmp_extract(archive=archive_path,
@@ -496,14 +492,10 @@ class HgArchiveBundle20Loader(HgBundle20Loader):
 
         repo_name = os.listdir(self.temp_dir)[0]
         directory = os.path.join(self.temp_dir, repo_name)
-        try:
-            super().prepare(origin_url=origin_url,
-                            visit_date=visit_date, directory=directory)
-        except Exception:
-            self.cleanup()
-            raise
+        super().prepare(origin_url=origin_url,
+                        visit_date=visit_date, directory=directory)
 
     def cleanup(self):
-        if os.path.exists(self.temp_dir):
+        if self.temp_dir and os.path.exists(self.temp_dir):
             rmtree(self.temp_dir)
         super().cleanup()
